fix: Ensure normalized image is of float32 type and apply ReLU in main function

+++ src/image.py
@@ -26,6 +26,6 @@ class Image:
             normalized = (image_float - img_min) / (img_max - img_min)
             normalized = 1.0 - normalized
         else:
-            normalized = np.zeros_like(image_float)
+            normalized = np.zeros_like(image_float, dtype=np.float32)
 
         return normalized
